finished implementing new logger along with some preparation for bigger rewrites
-finished implementing new logging system that is cross platform and less weird overall lmfao like why are we using arrays in cpp and functions like malloc, memcpy and free go back to C u fking purest this is C++ why do C programmers keep trying to use c++ if they're just gonna program it like C just use C assholes wtf

-removed weird c style strings and replaced with proper cpp style strings

- started replacing macros with constexpr when applicable

- started updating copyright in files

- more formatting updates according to the style i like and also replacing bool operators with actual words because human readable code is supposed to usewords thats the whole fucking point of a compiler wtf

everything compiles but the library is 1000% broken, i started preparations for replacing the weird callback system with a more sensible functional approach,

because i believe a library is supposed to give functionality to whatever program its used in and should expose everything thats going on opposed to a black box that just is "just trust me bro im doing what it says on the tin" kinda approach.

each function should be well defined in what it does, and should give more fine grained control to allow the end user to incorporate it into a custom control flow that makes sense for whatever application the library is being used in.

It is just more in the spirit of helping end users do what they want without putting non-sensical restrictions on how they want to achieve it. you should be free to code however you want, regardless of how fucking stupid it may be uwu